Update order management

+++ frontend/admin-console/src/data/dummy.js
@@ -1,6 +1,3 @@
-/* eslint-disable no-trailing-spaces */
-/* eslint-disable comma-dangle */
-/* eslint-disable quotes */
 import { t } from "i18next";
 import React from "react";
 import { AiOutlineCalendar, AiOutlineShoppingCart } from "react-icons/ai";
@@ -50,6 +47,7 @@ import organizationServices from "../api/services/organizationServices";
 import { has } from "lodash";
 import { ROLE, AccountStatus } from "../utils/constants";
 import { convertMongodbTimeToString } from "../utils/utils";
+import { setEventId, setOpenModal } from "../redux/slices/eventSlice";
 const { approveOrganizer, refuseOrganizer } = organizationServices;
 const { deleteEvent } = eventServices;
 export const gridOrderImage = (props) => (
@@ -71,187 +69,7 @@ export const gridOrderStatus = (props) => (
     {props.Status}
   </button>
 );
-const gridEventBackground = (props) => (
-  <div className="flex items-center gap-2">
-    <img
-      className="w-20 h-10 object-cover rounded-md"
-      src={props.background}
-      alt="employee"
-    />
-  </div>
-);
-const gridEventStatus = (props) => (
-  <div className="flex items-center gap-2">
-    {props.status === "event.available" ? (
-      <span className="bg-green-500 text-white p-1 text-xs rounded-md">
-        {t("event.status.available")}
-      </span>
-    ) : props.status === "event.completed" ? (
-      <span className="bg-yellow-500 text-white p-1 text-xs rounded-md">
-        {t("event.status.completed")}
-      </span>
-    ) : (
-      <span className="bg-red-600 text-white p-1 text-xs rounded-md">
-        {t("event.status.soldout")}
-      </span>
-    )}
-  </div>
-);
-const gridAccountStatus = (props) => (
-  <div className="flex items-center gap-2">
-    {props.organization.status !== "DISABLED" ? (
-      <span className="bg-green-500 text-white p-1 text-xs rounded-md">
-        {t("account.accepted")}
-      </span>
-    ) : (
-      <span className="bg-red-600 text-white p-1 text-xs rounded-md">
-        {t("account.disabled")}
-      </span>
-    )}
-  </div>
-);
-const gridEventQuantity = (props) => (
-  <div className="flex items-center justify-center">
-    <span>
-      {has(props, "organization") ? props.organization.eventList.length : null}
-    </span>
-  </div>
-);
-const gridCategoryName = (props) => (
-  <div className="flex items-center">
-    <span>{t(props.name)}</span>
-  </div>
-);
-const gridEventModify = (props) => (
-  <div className="flex items-center gap-2">
-    <RiEditFill
-      className="text-primary text-xl cursor-pointer"
-      onClick={() => window.location.replace(`/event/update/${props.id}`)}
-    />
-    <FaTrashAlt
-      className="text-primary text-xl cursor-pointer"
-      onClick={() => {
-        AlertQuestion({
-          title: t("popup.event.delete"),
-          callback: async () => {
-            const response = await deleteEvent(
-              props.id,
-              store.getState().account.userInfo.id
-            );
-            if (response.status === 200) {
-              AlertPopup({
-                title: t("popup.event.delete-success"),
-                timer: 5000,
-              });
-            }
-          },
-        });
-      }}
-    />
-  </div>
-);
-const gridAccountOption = (props) => (
-  <div className="flex items-center">
-    {props.organization.status === "DISABLED" ? (
-      <div className="flex items-center justify-center gap-x-2">
-        <button
-          className="text-white font-semibold bg-green-600 px-2 py-1 flex items-center gap-x-1 rounded-sm"
-          onClick={() => {
-            AlertQuestion({
-              title: t("popup.account.approve"),
-              callback: async () => {
-                const response = await approveOrganizer({
-                  email: props.organization.email,
-                });
-                if (response.status === 200) {
-                  AlertPopup({
-                    title: t("popup.account.approve-success"),
-                    timer: 5000,
-                  });
-                } else {
-                  AlertErrorPopup({
-                    title: t("popup.account.approve-fail"),
-                    timer: 5000,
-                  });
-                }
-              },
-            });
-          }}
-        >
-          <BsCheckLg />
-          <span>{t("account.approve")}</span>
-        </button>
-        <button
-          className="text-white font-semibold bg-red-600 px-2 py-1 flex items-center gap-x-1 rounded-sm"
-          onClick={() => {
-            AlertQuestion({
-              title: t("popup.account.refuse"),
-              callback: async () => {
-                const response = await refuseOrganizer({
-                  email: props.organization.email,
-                });
-                if (response.status === 200) {
-                  AlertPopup({
-                    title: t("popup.account.refuse-success"),
-                    timer: 5000,
-                  });
-                } else {
-                  AlertErrorPopup({
-                    title: t("popup.account.refuse-fail"),
-                    timer: 5000,
-                  });
-                }
-              },
-            });
-          }}
-        >
-          <BsXLg />
-          <span>{t("account.disabled")}</span>
-        </button>
-      </div>
-    ) : null}
-  </div>
-);
-const gridEventCategory = (props) => (
-  <div className="flex items-center gap-2">
-    {props.eventCategoryList.map((category) => (
-      <span className="bg-transparent border-2 border-gray-500 text-gray-500 px-1 py-1 rounded-md mr-1 text-xs">
-        {t(category.name)}
-      </span>
-    ))}
-  </div>
-);
-const gridEventDate = (props) => (
-  <div className="flex items-center gap-2">
-    {props.startingDate === props.endingDate ? (
-      <span>{props.startingDate}</span>
-    ) : (
-      <span>
-        {props.startingDate} ~ {props.endingDate}
-      </span>
-    )}
-  </div>
-);
-const customerGridImage = (props) => (
-  <div className="image flex gap-4">
-    <img
-      className="rounded-md w-10 h-10"
-      src={props.CustomerImage}
-      alt="employee"
-    />
-    <div>
-      <p>{props.CustomerName}</p>
-      <p>{props.CustomerEmail}</p>
-    </div>
-  </div>
-);
 
-const customerGridStatus = (props) => (
-  <div className="flex gap-2 justify-center items-center text-gray-700 capitalize">
-    <p style={{ background: props.StatusBg }} className="rounded-md h-3 w-3" />
-    <p>{props.Status}</p>
-  </div>
-);
 export const areaPrimaryXAxis = {
   valueType: "DateTime",
   labelFormat: "y",
@@ -573,16 +391,18 @@ export const eventColumns = [
           onClick={() => {
             AlertQuestion({
               title: t("popup.event.delete"),
-              callback: async () => {
-                const response = await deleteEvent(
-                  record.id,
-                  store.getState().account.userInfo.id
-                );
-                if (response.status === 200) {
-                  AlertPopup({
-                    title: t("popup.event.delete-success"),
-                    timer: 5000,
-                  });
+              callback: async (result) => {
+                if (result.isConfirmed) {
+                  const response = await deleteEvent(
+                    record.id,
+                    store.getState().account.userInfo.id
+                  );
+                  if (response.status === 200) {
+                    AlertPopup({
+                      title: t("popup.event.delete-success"),
+                      timer: 5000,
+                    });
+                  }
                 }
               },
             });
@@ -655,7 +475,10 @@ export const orderColumns = [
     render: (_, record) => (
       <div
         className="text-[#1f3e82] font-medium flex items-center gap-x-2 cursor-pointer"
-        onClick={() => window.location.replace(`/orders/${record.id}`)}
+        onClick={() => {
+          store.dispatch(setEventId(record.id));
+          store.dispatch(setOpenModal(true));
+        }}
       >
         <span>{t("view-detail")}</span> <BsFillEyeFill />{" "}
       </div>
@@ -687,16 +510,18 @@ export const categoryColumns = [
           onClick={() => {
             AlertQuestion({
               title: t("popup.event.delete"),
-              callback: async () => {
-                const response = await deleteEvent(
-                  record.id,
-                  store.getState().account.userInfo.id
-                );
-                if (response.status === 200) {
-                  AlertPopup({
-                    title: t("popup.event.delete-success"),
-                    timer: 5000,
-                  });
+              callback: async (result) => {
+                if (result.isConfirmed) {
+                  const response = await deleteEvent(
+                    record.id,
+                    store.getState().account.userInfo.id
+                  );
+                  if (response.status === 200) {
+                    AlertPopup({
+                      title: t("popup.event.delete-success"),
+                      timer: 5000,
+                    });
+                  }
                 }
               },
             });
@@ -878,20 +703,22 @@ export const pendingAccountsColumns = [
           onClick={() => {
             AlertQuestion({
               title: t("popup.account.approve"),
-              callback: async () => {
-                const response = await approveOrganizer({
-                  email: record.email,
-                });
-                if (response.status === 200) {
-                  AlertPopup({
-                    title: t("popup.account.approve-success"),
-                    timer: 5000,
+              callback: async (result) => {
+                if (result.isConfirmed) {
+                  const response = await approveOrganizer({
+                    email: record.email,
                   });
-                } else {
-                  AlertErrorPopup({
-                    title: t("popup.account.approve-fail"),
-                    timer: 5000,
-                  });
+                  if (response.status === 200) {
+                    AlertPopup({
+                      title: t("popup.account.approve-success"),
+                      timer: 5000,
+                    });
+                  } else {
+                    AlertErrorPopup({
+                      title: t("popup.account.approve-fail"),
+                      timer: 5000,
+                    });
+                  }
                 }
               },
             });
@@ -905,20 +732,22 @@ export const pendingAccountsColumns = [
           onClick={() => {
             AlertQuestion({
               title: t("popup.account.refuse"),
-              callback: async () => {
-                const response = await refuseOrganizer({
-                  email: record.email,
-                });
-                if (response.status === 200) {
-                  AlertPopup({
-                    title: t("popup.account.refuse-success"),
-                    timer: 5000,
+              callback: async (result) => {
+                if (result.isConfirmed) {
+                  const response = await refuseOrganizer({
+                    email: record.email,
                   });
-                } else {
-                  AlertErrorPopup({
-                    title: t("popup.account.refuse-fail"),
-                    timer: 5000,
-                  });
+                  if (response.status === 200) {
+                    AlertPopup({
+                      title: t("popup.account.refuse-success"),
+                      timer: 5000,
+                    });
+                  } else {
+                    AlertErrorPopup({
+                      title: t("popup.account.refuse-fail"),
+                      timer: 5000,
+                    });
+                  }
                 }
               },
             });
